+++ configs/benchmarks/classification/_base_/models/vit_base_deit_p16.py
@@ -1,0 +1,19 @@
+# model settings
+model = dict(
+    type='MixUpClassification',
+    pretrained=None,
+    alpha=[0.8, 1.0,],  # deit setting
+    mix_mode=["mixup", "cutmix",],
+    mix_args=dict(),
+    backbone=dict(
+        type='VisionTransformer',
+        arch='base',
+        img_size=224, patch_size=16,
+        drop_path_rate=0.1,
+    ),
+    head=dict(
+        type='VisionTransformerClsHead',  # mixup CE + label smooth
+        loss=dict(type='LabelSmoothLoss',
+            label_smooth_val=0.1, num_classes=1000, mode='original', loss_weight=1.0),
+        in_channels=768, num_classes=1000)
+)