+++ openmixup/utils/__init__.py
@@ -3,7 +3,7 @@ from .collect import nondist_forward_collect, dist_forward_collect
 from .collect_env import collect_env
 from .config_tools import ConfigGenerator, traverse_replace
 from .dist_utils import (allreduce_grads, allreduce_params, get_dist_info,
-                         init_dist, master_only)
+                         init_dist, master_only, sync_random_seed)
 from .fp16_utils import LossScaler, auto_fp16, force_fp32, wrap_fp16_model
 from .flops_counter import get_model_complexity_info
 from .logger import get_root_logger, print_log
@@ -16,7 +16,7 @@ from .setup_env import setup_multi_processes
 __all__ = [
     'AliasMethod', 'nondist_forward_collect', 'dist_forward_collect',
     'collect_env', 'ConfigGenerator', 'traverse_replace',
-    'allreduce_grads', 'allreduce_params', 'get_dist_info', 'init_dist', 'master_only',
+    'allreduce_grads', 'allreduce_params', 'get_dist_info', 'init_dist', 'master_only', 'sync_random_seed',
     'LossScaler', 'auto_fp16', 'force_fp32', 'wrap_fp16_model',
     'get_model_complexity_info', 'get_root_logger', 'print_log',
     'find_latest_checkpoint', 'multi_apply', 'tensor2imgs', 'unmap',